backend/requirements.txt
fastapi>=0.104
uvicorn[standard]>=0.24
websockets>=12.0
pydantic>=2.5
python-jose[cryptography]>=3.3
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
motor>=3.3
pymongo>=4.6
numpy>=1.26
scipy>=1.11
scikit-learn>=1.3
pandas>=2.1
joblib>=1.3
python-dotenv>=1.0
httpx>=0.25
